Better handling of NSS core affinity
My understanding is the name of the game is to get as much as possible off CPU0.

This update does 3 things:

FIrst of all, it handles both NSS cores. The previous version only handled the first NSS core. On my 3 ipq807x machines, the 2nd NSS core, nss_queue0, still had quite a chunk of interrupts. The set_affinity function has been updated to take a param specifying which occurrence of the irq name we are pinning.

2nd of all, I have updated the logic as follows:

NSS CORE 1 = CPU1 CPU2 CPU3 CPU0
NSS CORE 2 = CPU2 CPU3 CPU0 CPU1

The reason being that nss_queue0 is always hit the hardest, so this will move a very significant number of interrupts off CPU0.

3rd of all I've added a logger call so we should now see some basic logging where things are pinned to in logread.

Now the caveats:

a) I have both package bash and full gawk installed. This needs to be tested on a setup that has neither.

b) This assumes that things work as they should. What I mean by that is for example, having an IRQ _not_ on CPU0 does not impose a penalty in and of it self. But thats all beyond my pay grade ;)