Update quectel.sh
On my Arcadyan AW1000 the built in LTE Modem (EG12-EA) does not make a connection because either this script or the quectel-cm utility doesn't enable the radio. On a fresh boot the radio is in low power mode, at least on my particular device.

Also fixed a couple sections where "ip" should be "ipv4". The modem would connect but not create the virtual _4 device and set up the gateway/routes.